working logic for reading csv

+++ src/main/scala/sparktraining/model/AirLineDelay.scala
@@ -1,3 +1,3 @@
 package sparktraining.model
 
-case class AirLineDelay(sourceAirport: String, destinationAirport: String, delay: Long)
+case class AirLineDelay(sourceAirport: String, destinationAirport: String, delay: Double)
